+++ fiddles/angular2-cli/fiddle-0001-GenerateRoute/e2e/app.po.ts
@@ -1,9 +1,0 @@
-export class GenerateRoutePage {
-  navigateTo() {
-    return browser.get('/');
-  }
-
-  getParagraphText() {
-    return element(by.css('generate-route-app h1')).getText();
-  }
-}