+++ fiddles/angular2-cli/fiddle-0001-GenerateRoute/src/app/index.ts
@@ -1,2 +1,0 @@
-export {environment} from './environment';
-export {GenerateRouteAppComponent} from './generate-route.component';