+++ fiddles/angular2-cli/fiddle-0001-GenerateRoute/src/main.ts
@@ -1,9 +1,0 @@
-import { bootstrap } from '@angular/platform-browser-dynamic';
-import { enableProdMode } from '@angular/core';
-import { GenerateRouteAppComponent, environment } from './app/';
-
-if (environment.production) {
-  enableProdMode();
-}
-
-bootstrap(GenerateRouteAppComponent);