+++ fiddles/angular2-cli/fiddle-0003-ComponentResolver/e2e/app.po.ts
@@ -1,9 +1,0 @@
-export class Fiddle0038ComponentResolverPage {
-  navigateTo() {
-    return browser.get('/');
-  }
-
-  getParagraphText() {
-    return element(by.css('fiddle-0038-component-resolver-app h1')).getText();
-  }
-}