+++ scripts/bin/setup/mac/_shrinkwrap.sh
@@ -1,26 +1,45 @@
 #!/usr/bin/env bash
 # ---------------------------------------------------------------------------------------------------|
-#  School / Organization   : bradyhouse.io___________________________________________________________|
+#  Repo                    : https://github.com/bradyhouse/house_____________________________________|
 #  Specification           : N/A_____________________________________________________________________|
 #  Specification Path      : N/A_____________________________________________________________________|
 #  Author                  : brady house_____________________________________________________________|
-#  Create date             : 04/12/2016______________________________________________________________|
-#  Description             : MASTER ENVIRONMENT VARIABLE SOURCE______________________________________|
+#  Create date             : 05/26/2018______________________________________________________________|
+#  Description             : MASTER SHRINKWRAP SETUP SCRIPT FOR DARWIN_______________________________|
 #  Command line Arguments  : N/A_____________________________________________________________________|
 # ---------------------------------------------------------------------------------------------------|
 #  Revision History::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::|
 # ---------------------------------------------------------------------------------------------------|
-# Baseline Ver - See CHANGELOG @ 201703100420
-# 01/24/2018 - See CHANGELOG @ aurelia-dependencies-update
-# 05/26/2018 - See CHANGELOG @ 230_update_and_shrinkwrap
+# Baseline Ver - CHANGELOG @ 230_update_and_shrinkwrap
 # ---------------------------------------------------------------------------------------------------|
 
-export HOME_ROOT=$(mapHomePath;);
-export GITHUB_PUBLISH_REPO="bradyhouse.github.io";
-export GITHUB_ROOT="${HOME_ROOT}/github";
-export BUILD_NUM="201801240420";
-export BASH_PROFILE="${HOME_ROOT}/.bash_profile";
-export OS=$(getOS;);
-export NVM_VERSION="v8.9.4";
-export NG_TYPESCRIPT_VER="2.7.2";
-export AUTHOR="bradyhouse@gmail.com";
+
+function installShrinkwrap() {
+  groupLog "iinstallShrinkwrap";
+  installed=$(isInstalled "npm";);
+  if [[ "${installed}" == "true" ]]
+  then
+    npm install -g shrinkwrap;
+  else
+    return 1;
+  fi
+}
+
+
+function install() {
+  groupLog "install";
+  #try
+  (
+    installShrinkwrap || exit $?;
+  )
+  #catch
+  _rc=$?
+  case ${_rc} in
+    0)  echo "Done. shrinkwrap installed successfully."
+        ;;
+    *)  echo "foo bar! Something went wrong."
+        ;;
+  esac
+  #finally
+  exit ${_rc}
+}