+++ scripts/fiddle-build.sh
@@ -12,6 +12,7 @@
 # ---------------------------------------------------------------------------------------------------|
 # 03/10/2017 - Baseline Ver.
 # 01/24/2018 - See CHANGELOG @ aurelia-dependencies-update
+# 05/26/2018 - See CHANGELOG @ 230_update_and_shrinkwrap
 # ---------------------------------------------------------------------------------------------------|
 
 this=$0;
@@ -30,7 +31,6 @@ source bin/_env.sh;
 
     case ${_type} in
         'angular2-cli')
-            source bin/angular2-cli/.ngrc;
             source bin/angular2-cli/_build.sh;
             build ${_fiddle} || exit 87;
             ;;