+++ CHANGELOG.md
@@ -1100,8 +1100,6 @@ Alternately, the _Daily Diary of my Programming Dreams_.
     6.  Added [fiddles/angular2/fiddle-0043-ComponentRouting](fiddles/angular2/fiddle-0043-ComponentRouting) ~ [Issue 43](https://github.com/bradyhouse/house/issues/43)
 
 * **Angular2 CLI**
-    1.  Added [fiddles/angular2-cli/fiddle-0001-GenerateRoute](fiddles/angular2-cli/fiddle-0001-GenerateRoute) ~ [Issue 22](https://github.com/bradyhouse/house/issues/22)
-    2.  Added [fiddles/angular2-cli/fiddle-0002-GeneratePipe](fiddles/angular2-cli/fiddle-0002-GeneratePipe) ~ [Issue 27](https://github.com/bradyhouse/house/issues/27)
 
 * **Node**
     1.  Added [fiddles/node/fiddle-0003-ChildProcess](fiddles/node/fiddle-0003-ChildProcess) ~ [Issue 12](https://github.com/bradyhouse/house/issues/12)
@@ -1736,3 +1734,39 @@ Alternately, the _Daily Diary of my Programming Dreams_.
 * **Python**
 
  1. Added [fiddles/python/fiddle-0010-HttpRequest](fiddles/python/fiddle-0010-HttpRequest) ~ [Issue 228](https://github.com/bradyhouse/house/issues/228)
+
+
+### 230_update_and_shrinkwrap ~ [Issue #239](https://github.com/bradyhouse/house/issues/230)
+
+* **fiddle.sh (Enhancement)**
+  
+  1.  Updated the node version to v8.9.4
+    * Reflected changes:
+      1.  Updated [scripts/bin/_env.sh](scripts/bin/_env.sh)
+      2.  Deleted [scripts/bin/angular2-cli/.ngrc](scripts/bin/angular2-cli/.ngrc)
+          * Updated [scripts/fiddle-start.sh](scripts/fiddle-start.sh)
+          * Updated [scripts/fiddle-build.sh](scripts/fiddle-build.sh)
+          * Updated [scripts/fiddle-create.sh](scripts/fiddle-create.sh)
+      
+  2.  Added npm-check-updates setup support
+    * Reflected changes:
+      1.  Added [scripts/bin/setup/mac/_ncu.sh](scripts/bin/setup/mac/_ncu.sh)
+      2.  Updated [scripts/bin/setup/_setup.sh](scripts/bin/setup/_setup.sh)
+      3.  Updated [scripts/bin/_types.sh](scripts/bin/_types.sh)
+   
+  3.  Added shrinkwrap setup support
+    * Reflected changes:
+      1.  Added [scripts/bin/setup/mac/_shrinkwrap.sh](scripts/bin/setup/mac/_ncu.sh)
+      2.  Updated [scripts/bin/setup/_setup.sh](scripts/bin/setup/_setup.sh)
+      3.  Updated [scripts/bin/_types.sh](scripts/bin/_types.sh)
+  
+  4.  Added update command support
+    * Reflected change:
+      1.  Added [scripts/fiddle-update.sh](scripts/fiddle-update.sh)
+      2.  Updated [scripts/fiddle.sh](scripts/fiddle.sh)
+      3.  Updated [scripts/bin/_types.sh](scripts/bin/_types.sh)
+      4.  Updated [scripts/bin/angular2-cli/_start.sh](scripts/bin/angular2-cli/_start.sh)
+      5.  Added [scripts/bin/angular2-cli/_update.sh](scripts/bin/angular2-cli/_update.sh)
+      6.  Added [fiddles/angular2-cli/update.md](fiddles/angular2-cli/update.md)
+      7.  Updated [fiddles/angular2-cli/readme.adoc](fiddles/angular2-cli/readme.adoc)
+      8.  Added [scripts/bin/aurelia/_update.sh](scripts/bin/aurelia/_update.sh)